- Fix compile error "error: conversion from ‘int’ to ‘quint16’ {aka ‘short unsigned int’} may change value"

+++ src/o0baseauth.cpp
@@ -110,7 +110,7 @@ int O0BaseAuth::localPort() {
 
 void O0BaseAuth::setLocalPort(int value) {
     qDebug() << "O0BaseAuth::setLocalPort:" << value;
-    localPort_ = value;
+    localPort_ = (quint16)value;
     Q_EMIT localPortChanged();
 }
